@@ -5,7 +5,7 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^8.0.2",
+        "php": "^8.1",
         "bref/bref": "^1.7",
         "bref/laravel-bridge": "^1.2",
         "guzzlehttp/guzzle": "^7.2",
