@@ -2,25 +2,17 @@
 
 namespace App\Providers;
 
-use App\Tools\Slugify;
-use App\Tools\Tool;
-use App\Tools\Tools;
+use App\Services\Geo\GeoService;
+use App\Services\Geo\GoogleMaps;
 use Illuminate\Support\ServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
 {
-    public function boot(): void
+    public function register(): void
     {
-        $this->app->bind(Tools::class, function () {
-            $tools = [
-                'slugify' => new Tool(
-                    handler: new Slugify\Handler(),
-                    inertia: new Slugify\Inertia(config('tools.slugify.component')),
-                    validation: new Slugify\Validation() ,
-                ),
-            ];
-
-            return new Tools($tools);
-        });
+        $this->app->bind(
+            GeoService::class,
+            fn () => new GoogleMaps(config('services.google_maps.key'))
+        );
     }
 }
