@@ -1,19 +1,6 @@
 <?php
 
 return [
-
-    /*
-    |--------------------------------------------------------------------------
-    | Third Party Services
-    |--------------------------------------------------------------------------
-    |
-    | This file is for storing the credentials for third party services such
-    | as Mailgun, Postmark, AWS and more. This file provides the de facto
-    | location for this type of information, allowing packages to have
-    | a conventional file to locate the various service credentials.
-    |
-    */
-
     'mailgun' => [
         'domain' => env('MAILGUN_DOMAIN'),
         'secret' => env('MAILGUN_SECRET'),
@@ -34,5 +21,9 @@ return [
     'fathom' => [
         'domain' => env('FATHOM_DOMAIN', 'cdn.usefathom.com'),
         'site' => env('FATHOM_SITE'),
+    ],
+
+    'google_maps' => [
+        'key' => env('GOOGLE_MAPS_KEY'),
     ],
 ];
